:truck:移动功能，更符合插件用途
